+++ buildSrc/src/main/kotlin/software/amazon/app/platform/gradle/buildsrc/BasePlugin.kt
@@ -55,12 +55,18 @@ public open class BasePlugin : Plugin<Project> {
     // as well.
     val substitutions =
       mapOf(
+        "${APP_PLATFORM_GROUP}:di-common-public" to ":di-common:public",
         "${APP_PLATFORM_GROUP}:kotlin-inject-public" to ":kotlin-inject:public",
         "${APP_PLATFORM_GROUP}:kotlin-inject-contribute-impl-code-generators" to
           ":kotlin-inject-extensions:contribute:impl-code-generators",
         "${APP_PLATFORM_GROUP}:kotlin-inject-contribute-public" to
           ":kotlin-inject-extensions:contribute:public",
         "${APP_PLATFORM_GROUP}:kotlin-inject-impl" to ":kotlin-inject:impl",
+        "${APP_PLATFORM_GROUP}:ksp-common-public" to ":ksp-common:public",
+        "${APP_PLATFORM_GROUP}:metro-public" to ":metro:public",
+        "${APP_PLATFORM_GROUP}:metro-impl" to ":metro:impl",
+        "${APP_PLATFORM_GROUP}:metro-contribute-impl-code-generators" to
+          ":metro-extensions:contribute:impl-code-generators",
         "${APP_PLATFORM_GROUP}:presenter-public" to ":presenter:public",
         "${APP_PLATFORM_GROUP}:presenter-molecule-public" to ":presenter-molecule:public",
         "${APP_PLATFORM_GROUP}:presenter-molecule-impl" to ":presenter-molecule:impl",
@@ -110,6 +116,11 @@ public open class BasePlugin : Plugin<Project> {
       with(extensions.getByType(BaseYarnRootExtension::class.java)) {
         // Force the newer version due to https://github.com/amzn/app-platform/security/dependabot/5
         resolution("webpack-dev-server", "5.2.1")
+        // Force the newer version due to https://github.com/amzn/app-platform/security/dependabot/8
+        resolution("on-headers", "1.1.0")
+        // Force the newer version due to
+        // https://github.com/amzn/app-platform/security/dependabot/10
+        resolution("tmp", "0.2.4")
       }
     }
   }