+++ buildSrc/src/main/kotlin/software/amazon/app/platform/gradle/buildsrc/SdkPlugin.kt
@@ -24,11 +24,18 @@ internal object SdkPlugin {
     // :presenter:testing -> ${group}:presenter-testing:${version}
     val parent = requireNotNull(parent)
     val artifactId =
-      if (parent.name == "contribute" && parent.parent?.name == "kotlin-inject-extensions") {
-        // Change the artifact ID, because "contribute" alone is a weird name.
-        artifactId(libraryName = "kotlin-inject-contribute")
-      } else {
-        artifactId()
+      when {
+        parent.name == "contribute" && parent.parent?.name == "kotlin-inject-extensions" -> {
+          // Change the artifact ID, because "contribute" alone is a weird name.
+          artifactId(libraryName = "kotlin-inject-contribute")
+        }
+        parent.name == "contribute" && parent.parent?.name == "metro-extensions" -> {
+          // Change the artifact ID, because "contribute" alone is a weird name.
+          artifactId(libraryName = "metro-contribute")
+        }
+        else -> {
+          artifactId()
+        }
       }
     mavenPublish.coordinates(artifactId = artifactId)
     mavenPublish.pom { pom ->
@@ -57,7 +64,8 @@ internal object SdkPlugin {
 
     // These packages only contain generated code that is picked up by compiler plugins.
     // They don't need to be part of the API dumps.
-    apiValidation.ignoredPackages += setOf("app.platform.inject", "amazon.lastmile.inject")
+    apiValidation.ignoredPackages +=
+      setOf("app.platform.inject", "amazon.lastmile.inject", "metro.hints")
   }
 
   private fun Project.configureExplicitApi() {