+++ gradle-plugin/src/main/kotlin/software/amazon/app/platform/gradle/AppPlatformExtension.kt
@@ -13,6 +13,7 @@ import org.gradle.api.artifacts.dsl.DependencyHandler
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Property
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
+import org.jetbrains.kotlin.gradle.plugin.KotlinTarget
 import software.amazon.app.platform.gradle.ModuleStructurePlugin.Companion.testingSourceSets
 
 /**
@@ -20,6 +21,7 @@ import software.amazon.app.platform.gradle.ModuleStructurePlugin.Companion.testi
  * ```
  * appPlatform {
  *   enableKotlinInject true // false is the default
+ *   enableMetro true // false is the default
  *
  *   enableMoleculePresenters true // false is the default
  *   enableModuleStructure true // false is the default
@@ -51,6 +53,24 @@ constructor(objects: ObjectFactory, private val project: Project) {
   }
 
   internal fun isKotlinInjectEnabled(): Property<Boolean> = enableKotlinInject
+
+  private val enableMetro: Property<Boolean> =
+    objects.property(Boolean::class.java).convention(false)
+
+  /** Adds Metro as dependency. */
+  public fun enableMetro(enabled: Boolean) {
+    if (enabled == enableMetro.get()) return
+
+    enableMetro.set(enabled)
+    enableMetro.disallowChanges()
+
+    if (enabled) {
+      addPublicModuleDependencies(true)
+      project.enableMetro()
+    }
+  }
+
+  internal fun isMetroEnabled(): Property<Boolean> = enableMetro
 
   private val enableMoleculePresenters: Property<Boolean> =
     objects.property(Boolean::class.java).convention(false)
@@ -154,31 +174,29 @@ private fun Project.enableKotlinInject() {
   )
 
   fun DependencyHandler.addKspProcessorDependencies(kspConfigurationName: String) {
-    addProvider(
+    add(kspConfigurationName, "me.tatarka.inject:kotlin-inject-compiler-ksp:$KOTLIN_INJECT_VERSION")
+    add(
       kspConfigurationName,
-      provider { "me.tatarka.inject:kotlin-inject-compiler-ksp:" + KOTLIN_INJECT_VERSION },
+      "$APP_PLATFORM_GROUP:kotlin-inject-contribute-public:$APP_PLATFORM_VERSION",
     )
     add(
       kspConfigurationName,
-      "$APP_PLATFORM_GROUP:kotlin-inject-contribute-public:" + APP_PLATFORM_VERSION,
+      "$APP_PLATFORM_GROUP:kotlin-inject-contribute-impl-code-generators:$APP_PLATFORM_VERSION",
     )
     add(
       kspConfigurationName,
-      "$APP_PLATFORM_GROUP:kotlin-inject-contribute-impl-code-generators:" + APP_PLATFORM_VERSION,
-    )
-    add(
-      kspConfigurationName,
-      "software.amazon.lastmile.kotlin.inject.anvil:compiler:" + KOTLIN_INJECT_ANVIL_VERSION,
+      "software.amazon.lastmile.kotlin.inject.anvil:compiler:$KOTLIN_INJECT_ANVIL_VERSION",
     )
   }
 
   plugins.withId(PluginIds.KOTLIN_MULTIPLATFORM) {
     kmpExtension.sourceSets.getByName("commonMain").dependencies {
       implementation("me.tatarka.inject:kotlin-inject-runtime:$KOTLIN_INJECT_VERSION")
+      implementation("$APP_PLATFORM_GROUP:di-common-public:$APP_PLATFORM_VERSION")
       implementation("$APP_PLATFORM_GROUP:kotlin-inject-public:$APP_PLATFORM_VERSION")
-      implementation("$APP_PLATFORM_GROUP:kotlin-inject-contribute-public:" + APP_PLATFORM_VERSION)
+      implementation("$APP_PLATFORM_GROUP:kotlin-inject-contribute-public:$APP_PLATFORM_VERSION")
       implementation(
-        "software.amazon.lastmile.kotlin.inject.anvil:runtime:" + KOTLIN_INJECT_ANVIL_VERSION
+        "software.amazon.lastmile.kotlin.inject.anvil:runtime:$KOTLIN_INJECT_ANVIL_VERSION"
       )
       implementation(
         "software.amazon.lastmile.kotlin.inject.anvil:runtime-optional:" +
@@ -187,44 +205,69 @@ private fun Project.enableKotlinInject() {
     }
 
     kmpExtension.targets.configureEach { target ->
-      if (target.name != "metadata") {
-        dependencies.addKspProcessorDependencies("ksp${target.name.capitalize()}")
-        dependencies.addKspProcessorDependencies("ksp${target.name.capitalize()}Test")
-
-        if (target.platformType == KotlinPlatformType.androidJvm) {
-          target.compilations.configureEach { compilation ->
-            if (compilation.name == "debugAndroidTest") {
-              // This is the name of the configuration for instrumented tests in
-              // KMP projects.
-              dependencies.addKspProcessorDependencies("kspAndroidAndroidTest")
-            }
-          }
-        }
+      addKspDependenciesWhenConfigExists(target) { configName ->
+        dependencies.addKspProcessorDependencies(configName)
       }
     }
   }
 
   plugins.withIds(PluginIds.KOTLIN_ANDROID, PluginIds.KOTLIN_JVM) {
+    dependencies.add("implementation", "$APP_PLATFORM_GROUP:di-common-public:$APP_PLATFORM_VERSION")
     dependencies.add(
       "implementation",
       "$APP_PLATFORM_GROUP:kotlin-inject-public:$APP_PLATFORM_VERSION",
     )
     dependencies.add(
       "implementation",
-      "$APP_PLATFORM_GROUP:kotlin-inject-contribute-public:" + APP_PLATFORM_VERSION,
+      "$APP_PLATFORM_GROUP:kotlin-inject-contribute-public:$APP_PLATFORM_VERSION",
     )
     dependencies.add(
       "implementation",
-      "software.amazon.lastmile.kotlin.inject.anvil:runtime:" + KOTLIN_INJECT_ANVIL_VERSION,
+      "software.amazon.lastmile.kotlin.inject.anvil:runtime:$KOTLIN_INJECT_ANVIL_VERSION",
     )
     dependencies.add(
       "implementation",
-      "software.amazon.lastmile.kotlin.inject.anvil:runtime-optional:" + KOTLIN_INJECT_ANVIL_VERSION,
+      "software.amazon.lastmile.kotlin.inject.anvil:runtime-optional:$KOTLIN_INJECT_ANVIL_VERSION",
     )
     dependencies.add(
       "implementation",
-      "me.tatarka.inject:kotlin-inject-runtime:" + KOTLIN_INJECT_VERSION,
+      "me.tatarka.inject:kotlin-inject-runtime:$KOTLIN_INJECT_VERSION",
     )
+    dependencies.addKspProcessorDependencies("ksp")
+  }
+}
+
+private fun Project.enableMetro() {
+  plugins.apply(PluginIds.METRO)
+
+  // Enable KSP for our custom extensions.
+  plugins.apply(PluginIds.KSP)
+
+  fun DependencyHandler.addKspProcessorDependencies(kspConfigurationName: String) {
+    add(
+      kspConfigurationName,
+      "$APP_PLATFORM_GROUP:metro-contribute-impl-code-generators:$APP_PLATFORM_VERSION",
+    )
+  }
+
+  plugins.withId(PluginIds.KOTLIN_MULTIPLATFORM) {
+    kmpExtension.sourceSets.getByName("commonMain").dependencies {
+      implementation("$APP_PLATFORM_GROUP:di-common-public:$APP_PLATFORM_VERSION")
+      implementation("$APP_PLATFORM_GROUP:metro-public:$APP_PLATFORM_VERSION")
+
+      kmpExtension.targets.configureEach { target ->
+        addKspDependenciesWhenConfigExists(target) { configName ->
+          dependencies.addKspProcessorDependencies(configName)
+        }
+      }
+    }
+  }
+
+  plugins.withIds(PluginIds.KOTLIN_ANDROID, PluginIds.KOTLIN_JVM) {
+    dependencies.add("implementation", "$APP_PLATFORM_GROUP:di-common-public:$APP_PLATFORM_VERSION")
+
+    dependencies.add("implementation", "$APP_PLATFORM_GROUP:metro-public:$APP_PLATFORM_VERSION")
+
     dependencies.addKspProcessorDependencies("ksp")
   }
 }
@@ -311,6 +354,45 @@ private fun Project.enableComposeUi() {
         "androidTestImplementation",
         "$APP_PLATFORM_GROUP:robot-compose-multiplatform-public:$APP_PLATFORM_VERSION",
       )
+    }
+  }
+}
+
+private fun Project.addKspDependenciesWhenConfigExists(
+  target: KotlinTarget,
+  block: (String) -> Unit,
+) {
+  if (target.name != "metadata") {
+    target.compilations.configureEach { compilation ->
+      fun configExists(name: String): Boolean = configurations.any { it.name == name }
+
+      // The implementationConfigurationName name is
+      // 'iosSimulatorArm64CompilationImplementation', 'wasmJsTestCompileClasspath' or
+      // 'desktopCompileClasspath'.
+      //
+      // E.g. 'desktopCompileClasspath' with give use 'kspDesktop'
+      var configName =
+        "ksp" +
+          compilation.implementationConfigurationName.substringBefore("Compilation").capitalize()
+
+      if (!configExists(configName) && target.platformType == KotlinPlatformType.androidJvm) {
+        // Android has different naming for some reason.
+        //
+        // E.g. for instrumentation tests 'kspAndroidDebugAndroidTest' should actually be
+        // 'kspAndroidAndroidTestDebug', but we will use 'kspAndroidAndroidTest'.
+        //
+        // For unit tests 'kspAndroidDebugUnitTest' should actually be 'kspAndroidTestDebug',
+        // but we will use 'kspAndroidTest'.
+        when {
+          configName.endsWith("AndroidTest") -> configName = "kspAndroidAndroidTest"
+          configName.endsWith("UnitTest") -> configName = "kspAndroidTest"
+        }
+      }
+
+      // Check again if the config exists.
+      if (configExists(configName)) {
+        block(configName)
+      }
     }
   }
 }