+++ gradle-plugin/src/main/kotlin/software/amazon/app/platform/gradle/PluginIds.kt
@@ -2,6 +2,7 @@ package software.amazon.app.platform.gradle
 
 internal object PluginIds {
   const val ANDROID_APP = "com.android.application"
+  const val ANDROID_KMP_LIBRARY = "com.android.kotlin.multiplatform.library"
   const val ANDROID_LIBRARY = "com.android.library"
   const val COMPOSE_COMPILER = "org.jetbrains.kotlin.plugin.compose"
   const val COMPOSE_MULTIPLATFORM = "org.jetbrains.compose"
@@ -9,4 +10,5 @@ internal object PluginIds {
   const val KOTLIN_JVM = "org.jetbrains.kotlin.jvm"
   const val KOTLIN_ANDROID = "org.jetbrains.kotlin.android"
   const val KSP = "com.google.devtools.ksp"
+  const val METRO = "dev.zacsweers.metro"
 }